+++ src/gfx.h
@@ -9,7 +9,7 @@ typedef struct renderer_t {
 	Rectangle framebufferRect, displayRect;
 } renderer_t;
 
-void renderer_init(renderer_t *render, float screen_width, float screen_height,
-				   const framebuffer_t *fb);
+void renderer_init(renderer_t *render, uint8_t screen_width,
+				   uint8_t screen_height, const framebuffer_t *fb);
 
 void renderer_draw_framebuffer(renderer_t *render, const framebuffer_t *fb);
